int -> float

+++ Formular/src/org/formular/core/Input.java
@@ -1,6 +1,6 @@
 package org.formular.core;
 
 public interface Input {
-	public void setVal(int i);
-	public int getVal();
+	public void setVal(float i);
+	public float getVal();
 }
