+++ Formular/src/org/formular/core/Somme.java
@@ -9,7 +9,7 @@ public class Somme extends OperationComposition {
 	private static final long serialVersionUID = 1L;
 
 	@Override
-	public int result() {
+	public float result() {
 		if(right == null || left == null)
 			return 0;
 		return right.result() + left.result();
